feat: added QR popover component
