+++ packages/ui/src/components/card-buttons/index.tsx
@@ -1,17 +1,13 @@
 import clsx from 'clsx'
-import { QRCodeSVG } from 'qrcode.react'
 import React from 'react'
 import { defineMessages, useIntl } from 'react-intl'
 import { useParams } from 'react-router-dom'
 
 import { Box } from 'ui/src/components/box'
-import { CopyAddressButton } from 'ui/src/components/copy-address-button'
 import { QrCode2Icon, UpRight2Icon } from 'ui/src/components/icons'
-import { PopoverContent, PopoverPortal, PopoverRoot, PopoverTrigger } from 'ui/src/components/popover'
+import { QrPopOver } from 'ui/src/components/qr-popover'
 import { Button } from 'ui/src/components/router-button'
 import { ToolTip } from 'ui/src/components/tool-tip'
-import { Text } from 'ui/src/components/typography'
-import { getShortAddress } from 'ui/src/utils/string-utils'
 
 import * as styles from './styles.css'
 
@@ -25,8 +21,8 @@ const messages = defineMessages({
 		defaultMessage: 'Send',
 	},
 	address: {
-		id: 'f4sJ8p',
-		defaultMessage: 'Address QR',
+		id: 'hc47g1',
+		defaultMessage: 'Address QR code',
 	},
 })
 
@@ -48,41 +44,13 @@ export const CardButtons: React.FC<IProps> = props => {
 					<UpRight2Icon />
 				</Button>
 			</ToolTip>
-
-			<PopoverRoot>
-				<PopoverTrigger asChild>
-					<Box>
-						<ToolTip message={intl.formatMessage(messages.address)}>
-							<Button iconOnly rounded styleVariant="inverse" sizeVariant={{ mobile: 'medium', tablet: 'large' }}>
-								<QrCode2Icon />
-							</Button>
-						</ToolTip>
-					</Box>
-				</PopoverTrigger>
-				<PopoverPortal>
-					<PopoverContent align="center" sideOffset={10} style={{ maxWidth: '300px' }}>
-						<Box padding="medium" display="flex" flexDirection="column" gap="small">
-							<Box display="flex" alignItems="center" gap="small">
-								<Text size="xsmall" weight="medium">
-									{getShortAddress(accountId)}
-								</Text>
-								<CopyAddressButton
-									styleVariant="ghost"
-									sizeVariant="xsmall"
-									address={accountId}
-									iconOnly
-									rounded={false}
-									tickColor="white"
-									toolTipDisabled
-								/>
-							</Box>
-							<Box>
-								<QRCodeSVG value={accountId} size={180} className={styles.cardButtonsQrCode} />
-							</Box>
-						</Box>
-					</PopoverContent>
-				</PopoverPortal>
-			</PopoverRoot>
+			<QrPopOver address={accountId}>
+				<ToolTip message={intl.formatMessage(messages.address)}>
+					<Button iconOnly rounded styleVariant="inverse" sizeVariant={{ mobile: 'medium', tablet: 'large' }}>
+						<QrCode2Icon />
+					</Button>
+				</ToolTip>
+			</QrPopOver>
 		</Box>
 	)
 }
