+++ packages/ui/src/components/card-buttons/styles.css.ts
@@ -1,6 +1,6 @@
-import { globalStyle, style } from '@vanilla-extract/css'
+import { style } from '@vanilla-extract/css'
 
-import { darkMode, sprinkles } from 'ui/src/components/system/sprinkles.css'
+import { sprinkles } from 'ui/src/components/system/sprinkles.css'
 
 export const cardButtonsWrapper = style([
 	sprinkles({
@@ -10,26 +10,3 @@ export const cardButtonsWrapper = style([
 	}),
 	{},
 ])
-
-export const cardButtonsQrCode = style([
-	sprinkles({
-		position: 'relative',
-	}),
-	{},
-])
-
-globalStyle(`${cardButtonsQrCode} path:nth-child(1)`, {
-	fill: '#161718 !important',
-})
-
-globalStyle(`.${darkMode} ${cardButtonsQrCode} path:nth-child(1)`, {
-	fill: '#a6a6a6 !important',
-})
-
-globalStyle(`${cardButtonsQrCode} path:nth-child(2)`, {
-	fill: '#ffffff !important',
-})
-
-globalStyle(`.${darkMode} ${cardButtonsQrCode} path:nth-child(2)`, {
-	fill: '#161718 !important',
-})
